nuscenes-api/requirements.txt
fastapi
uvicorn[standard]
nuscenes-devkit
pydantic-settings
python-dotenv
Shapely>=2.0.0